requirements.txt

hydra-core
pyyaml
